# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 공통기능 25개 API
  version: 0.1.0
servers:
- url: /
paths:
...
  /api/admin/base-years:
    get:
      operationId: getBaseYears
      summary: 기준연도 관리 조회
      x-related-requirements:
      - REQ-045
      - REQ-046
      - REQ-047
      - REQ-048
      - REQ-105
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-149
      - REQ-166
      - REQ-167
      - REQ-181
      - REQ-194
      - REQ-198
      - REQ-218
      x-roles:
      - R09
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          enum:
          - 20
          - 50
          - 100
          default: 20
      - name: sort
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
    post:
      operationId: saveBaseYears
      summary: 기준연도 관리 저장/실행
      x-related-requirements:
      - REQ-045
      - REQ-046
      - REQ-047
      - REQ-048
      - REQ-105
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-120
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-149
      - REQ-166
      - REQ-167
      - REQ-181
      - REQ-194
      - REQ-198
      - REQ-218
      x-roles:
      - R09
      x-business-rules:
      - 기준연도 관리 원문 제약과 서버 권한을 적용한다
      - base_years 저장 전 필수값·기간·중복 조건을 검증한다
      x-side-effects:
      - base_years 레코드를 등록·수정하거나 실행 이력을 기록한다
      - audit_logs에 기준연도 관리 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 기준연도 관리 저장 -> base_years 후속 조회에 반영'
      - 'auth: R01이 기준연도 관리 저장 요청 -> 403 반환'
      - 'validation: 필수 식별자 누락 -> fields.id 오류'
      - 'business: 원문 불변/기간/중복 제약 위반 -> base_years 상태 변경 없음'
      - 'side-effect: 성공 처리 후 audit_logs에 기준연도 관리 이력 존재'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
  /api/admin/batch-executions:
    get:
      operationId: runBatch
      summary: 배치 실행 관리 조회
      x-related-requirements:
      - REQ-096
      - REQ-097
      - REQ-098
      - REQ-099
      - REQ-105
      - REQ-113
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-120
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-163
      - REQ-166
      - REQ-167
      - REQ-192
      - REQ-194
      - REQ-198
      x-roles:
      - R09
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          enum:
          - 20
          - 50
          - 100
          default: 20
      - name: sort
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
    post:
      operationId: runBatch
      summary: 배치 실행 관리 저장/실행
      x-related-requirements:
      - REQ-096
      - REQ-097
      - REQ-098
      - REQ-099
      - REQ-105
      - REQ-113
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-120
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-163
      - REQ-166
      - REQ-167
      - REQ-192
      - REQ-194
      - REQ-198
      x-roles:
      - R09
      x-business-rules:
      - 배치 실행 관리 원문 제약과 서버 권한을 적용한다
      - batch_executions 저장 전 필수값·기간·중복 조건을 검증한다
      x-side-effects:
      - batch_executions 레코드를 등록·수정하거나 실행 이력을 기록한다
      - audit_logs에 배치 실행 관리 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 배치 실행 관리 저장 -> batch_executions 후속 조회에 반영'
      - 'auth: R01이 배치 실행 관리 저장 요청 -> 403 반환'
      - 'validation: 필수 식별자 누락 -> fields.id 오류'
      - 'business: 원문 불변/기간/중복 제약 위반 -> batch_executions 상태 변경 없음'
      - 'side-effect: 성공 처리 후 audit_logs에 배치 실행 관리 이력 존재'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
  /api/admin/batch-executions/{executionId}/stop:
    post:
      operationId: stopBatch
      summary: 배치 중지
      x-related-requirements:
      - REQ-096
      - REQ-097
      - REQ-098
      - REQ-099
      - REQ-113
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-120
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-163
      - REQ-166
      - REQ-167
      - REQ-192
      - REQ-198
      x-roles:
      - R09
      x-business-rules:
      - 배치 중지 권한과 원문 제약을 검증한다
      x-side-effects:
      - 배치 중지 처리 결과 또는 감사 이력을 기록한다
      x-required-tests:
      - 'happy: 배치 중지 요청 -> 후속 조회로 결과 확인'
      - 'auth: R01 요청 -> 403 반환'
      - 'validation: path id 누락/오류 -> ApiError 반환'
      - 'business: 원문 차단 조건 위반 -> 대상 상태 유지'
      - 'side-effect: 성공 시 audit_logs 또는 이력 테이블에 기록 확인'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
  /api/admin/batch-executions/{executionId}/rerun:
    post:
      operationId: rerunBatch
      summary: 배치 재실행
      x-related-requirements:
      - REQ-096
      - REQ-097
      - REQ-098
      - REQ-099
      - REQ-113
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-120
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-163
      - REQ-166
      - REQ-167
      - REQ-192
      - REQ-198
      x-roles:
      - R09
      x-business-rules:
      - 배치 재실행 권한과 원문 제약을 검증한다
      x-side-effects:
      - 배치 재실행 처리 결과 또는 감사 이력을 기록한다
      x-required-tests:
      - 'happy: 배치 재실행 요청 -> 후속 조회로 결과 확인'
      - 'auth: R01 요청 -> 403 반환'
      - 'validation: path id 누락/오류 -> ApiError 반환'
      - 'business: 원문 차단 조건 위반 -> 대상 상태 유지'
      - 'side-effect: 성공 시 audit_logs 또는 이력 테이블에 기록 확인'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
    ApiResponse:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - true
        data: {}
        meta:
          type: object
    ApiError:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
        meta:
          type: object
...
security:
- sessionCookie: []
...